3.12 implement folds with the opposite-handed fold
i.e. foldLeft using foldRight, and foldRight using foldLeft.

I did both of these by first reversing the list with a fold, and then
doing the opposite-handed fold on the reversed list.

Reversing is easy with foldLeft, gnarly with foldRight. Uses a recursive
function that will be very slow. Doubt this is the solution the book had
in mind.